# docs/swagger.yaml
basePath: /
definitions:
  api.uploadFileResponse:
    properties:
      message:
        type: string
      url:
        items:
          type: string
        type: array
    type: object
host: localhost:8081
info:
  contact: {}
  description: This is a documentation for Segment3d App API
  title: Segment3d App API Documentation
  version: "1.0"
paths:
  /files/{path}:
    get:
      consumes:
      - application/json
      description: Retrieve file data from specified path within the server's storage
        directory.
      parameters:
      - description: Path including any folders and subfolders to the file
        in: path
        name: path
        required: true
        type: string
      - description: isReturnLink
        in: query
        name: isLink
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: File retrieved successfully
          schema:
            type: file
      summary: Get file
      tags:
      - file
  /thumbnail/{path}:
    get:
      consumes:
      - application/json
      description: Retrieve thumbnail from specified resource path
      parameters:
      - description: Path including any folders and subfolders to the file
        in: path
        name: path
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: File retrieved successfully
          schema:
            type: file
      summary: Get file
      tags:
      - file
  /upload:
    post:
      consumes:
      - multipart/form-data
      description: Uploads a file to the specified folder within the server's storage
        directory.
      parameters:
      - description: Folder where the file will be uploaded
        in: formData
        name: folder
        required: true
        type: string
      - collectionFormat: csv
        description: File(s) to upload
        in: formData
        items:
          type: file
        name: file
        required: true
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: Upload file success
          schema:
            $ref: '#/definitions/api.uploadFileResponse'
      summary: Upload file
      tags:
      - file
swagger: "2.0"